achanges in DTO

+++ src/main/java/com/ILPex/DTO/CourseProgressDTO.java
@@ -13,5 +13,5 @@ public class CourseProgressDTO {
     private Integer dayNumber;
     private Integer estimatedDuration;
     private Integer duration;
-   private Double percentageCompleted;
+   private Integer percentageCompleted;
 }
